==> ELSA/6_Physical_measures/1_weight.Rmd <==
---
title: "Weight"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("../../setup_iago.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:


* Short name: `weight`
* Variable label: `Weight of the participant in kg`
* Domain: `Physical measures `
* Value type: `Continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

* Comments: `Weight should be collected in kilograms. It should be determined whether the  variable was collected through a self reported measure or an athropometric test.`




<!-- ########################################################## --> 

# Data process






```{r, dataDirectory, echo=FALSE}
datafolder <- "../RData"
```

## England
### Wave 1 

There is no variable. 



### Wave 2

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `weight`
* Label: `Weight (kg) including unreliable measures`
* Type of variable: `countinous`
* Missings
    + `-1 = Not Applicable`


```{r assign2, echo=F}
opal.assign.table.tibble(o, 'ELSA_w2_weig','ELSA.wave_2_nurse_data_v2',
                         variables=list('weight'), missings = TRUE)
```

```{r local2, echo=F}
ELSA_w2_weig <- opal.execute(o,'ELSA_w2_weig')
Continuous_summary(var = ELSA_w2_weig$weight, missing_values = c(-1))
ggplot(ELSA_w2_weig, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency") + xlim(30,130)
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `recode -1 "Not Applicable" into 996`
* `NA into 999`

**R script:**

```{r harmo2}
ELSA_ds_w2 <- tibble(id=ELSA_w2_weig$id)
ELSA_ds_w2$weight <- car::recode(ELSA_w2_weig$weight, "NA=999; -1=996")
ELSA_ds_w2$weight <- labelled(ELSA_ds_w2$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=F}
Continuous_summary(var = ELSA_ds_w2$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ELSA_ds_w2, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(30,130)
```

#### Validation
It is going to check if weight and heigth values are correct by computing BMI with the formula: weight/height^2. 
```{r crosstabulation2, echo=F}
opal.assign.table.tibble(o, 'ELSA_w2_whb','ELSA.wave_2_nurse_data_v2',
                         variables=list('weight', 'htval', 'bmival'), missings = TRUE)
ELSA_w2_whb <- opal.execute(o,'ELSA_w2_whb')

compute_bmi <- function(Heigth, Weigth, miss, heigth_m = TRUE){
  Heigth <- car::recode(Heigth, "NA=999")
  Weigth <- car::recode(Weigth, "NA=999")
  
  if(heigth_m){
    bmi <- Weigth/(Heigth^2)
    
    bmi[Heigth==999] <- 999
    bmi[Heigth==miss[1]] <- Heigth[Heigth==miss[1]]
    
    bmi[Weigth==999] <- 999
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[1]]
  } else {
    bmi <- Weigth/((Heigth/100)^2)
    
    bmi[Heigth==999] <- 999
    bmi[Heigth==miss[1]] <- Heigth[Heigth==miss[1]]
    
    bmi[Weigth==999] <- 999
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[1]]
  }
  bmi
}

ELSA_w2_whb$bmi_c <- compute_bmi(Heigth = ELSA_w2_whb$htval, 
                        Weigth = ELSA_w2_whb$weight, miss = c(-1), 
                        heigth_m = FALSE)

ELSA_w2_whb$proff <- ELSA_w2_whb$bmival - ELSA_w2_whb$bmi_c

head(ELSA_w2_whb[which(ELSA_w2_whb$proff>2 | ELSA_w2_whb$proff<(-2)),])

bmi_na <- which(ELSA_w2_whb$bmival==(-1) & ELSA_w2_whb$htval!=(-1) & ELSA_w2_whb$weight!=(-1))
weig_na <- which(ELSA_w2_whb$bmival!=(-1) & ELSA_w2_whb$weight==(-1))
```

There are 32 individuals that the bmi value is different than the computed by the formula. Those discrepancies are because either bmi is missing (-1) or weight or height is missing (-1). Therefore, it is going to be used this values to complet the variables bmi, weight and height. 

```{r crosstabulation2_2, echo=F}
ELSA_ds_w2$weight[weig_na] <- ELSA_w2_whb$bmival[weig_na] * (ELSA_w2_whb$htval[weig_na]/100)^2
```

```{r importOPAL2, echo=F}
#val_labels(ELSA_ds_w2) <- NULL
#opal.assign.data(o, 'ELSA_ds_w2', ELSA_ds_w2)
#opal.symbol_import(o,'ELSA_ds_w2', project='_Athlos_Harm_Dataset')
```




```{r,  echo=FALSE}
weight <- ELSA_ds_w2
save(weight, file = paste0(datafolder, "/w2/weight.RData"))
```




### Wave 3

There is no variable.



### Wave 4

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `weight`
* Label: `Weight (kg) including unreliable measures`
* Type of variable: `countinous`
* Missings
    + `-1 = Not Applicable`

```{r assign4, echo=F}
opal.assign.table.tibble(o, 'ELSA_w4_weigh','ELSA.wave_4_nurse_data',
                         variables=list('weight'), missings = TRUE)
```

```{r local4, echo=F}
ELSA_w4_weigh <- opal.execute(o,'ELSA_w4_weigh')
Continuous_summary(var = ELSA_w4_weigh$weight, missing_values = c(-1))
ggplot(ELSA_w4_weigh, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency") + xlim(30,130)
```

There is a value 1 that it will be consider as a mistake intead of -1, therefore it is going to be recoded as 994. There is a value 10.4 that is going to consider a missing, therefore is going to recode as 999. 

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `recode -1 "Not Applicable" into 996`
* `NA into 999`

**R script:**

```{r harmo4}
ELSA_ds_w4 <- tibble(id=ELSA_w4_weigh$id)
ELSA_ds_w4$weight <- car::recode(ELSA_w4_weigh$weight, "NA=999; -1=996; 1=994; 10.4=999")
ELSA_ds_w4$weight <- labelled(ELSA_ds_w4$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript4, echo=F}
Continuous_summary(var = ELSA_ds_w4$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ELSA_ds_w4, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency") + xlim(30,130)
```

#### Validation
It is going to check if weight and heigth values are correct by computing BMI with the formula: weight/height^2. 
```{r crosstabulation4, echo=F}
opal.assign.table.tibble(o, 'ELSA_w4_whb','ELSA.wave_4_nurse_data',
                         variables=list('weight', 'htval', 'bmival'), missings = TRUE)
ELSA_w4_whb <- opal.execute(o,'ELSA_w4_whb')

compute_bmi <- function(Heigth, Weigth, miss, heigth_m = TRUE){
  Heigth <- car::recode(Heigth, "NA=999")
  Weigth <- car::recode(Weigth, "NA=999")
  
  if(heigth_m){
    bmi <- Weigth/(Heigth^2)
    
    bmi[Heigth==999] <- 999
    bmi[Heigth==miss[1]] <- Heigth[Heigth==miss[1]]
    
    bmi[Weigth==999] <- 999
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[1]]
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[2]]
  } else {
    bmi <- Weigth/((Heigth/100)^2)
    
    bmi[Heigth==999] <- 999
    bmi[Heigth==miss[1]] <- Heigth[Heigth==miss[1]]
    
    bmi[Weigth==999] <- 999
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[1]]
    bmi[Weigth==miss[2]] <- Weigth[Weigth==miss[2]]
  }
  bmi
}

ELSA_w4_whb$bmi_c <- compute_bmi(Heigth = ELSA_w4_whb$htval, 
                        Weigth = ELSA_w4_whb$weight, miss = c(-1,1), 
                        heigth_m = FALSE)

ELSA_w4_whb$proff <- ELSA_w4_whb$bmival - ELSA_w4_whb$bmi_c

head(ELSA_w4_whb[which(ELSA_w4_whb$proff>2 | ELSA_w4_whb$proff<(-2)),])

bmi_na <- which(ELSA_w4_whb$bmival==(-1) & ELSA_w4_whb$htval!=(-1) & ELSA_w4_whb$weight!=(-1) & ELSA_w4_whb$weight!=(1))
weig_na <- which(ELSA_w4_whb$bmival!=(-1) & ELSA_w4_whb$weight==(-1))
```

There are 49 individuals that the bmi value is different than the computed by the formula. Those discrepancies are because either bmi is missing (-1) or weight or height is missing (-1). Therefore, it is going to be used this values to complet the variables bmi, weight and height. 

```{r crosstabulation4_2, echo=F}
ELSA_ds_w4$weight[weig_na] <- ELSA_w4_whb$bmival[weig_na] * (ELSA_w4_whb$htval[weig_na]/100)^2

```

```{r importOPAL4, echo=F}
#val_labels(ELSA_ds_w4) <- NULL
#opal.assign.data(o, 'ELSA_ds_w4', ELSA_ds_w4)
#opal.symbol_import(o,'ELSA_ds_w4', project='_Athlos_Harm_Dataset')
```



```{r,  echo=FALSE}
weight <- ELSA_ds_w4
save(weight, file = paste0(datafolder, "/w4/weight.RData"))
```




### Wave 5

There is no varibale.


### Wave 6

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `weight`
* Label: `Weight (kg) including unreliable measures`
* Type of variable: `countinous`
* Missings
    + `-1 = Not Applicable`

```{r assign6, echo=F}
opal.assign.table.tibble(o, 'ELSA_w6_weig','ELSA.wave_6_elsa_nurse_data_v2',
                         variables=list('weight'), missings = TRUE)
```

```{r local6, echo=F}
ELSA_w6_weig <- opal.execute(o,'ELSA_w6_weig')
Continuous_summary(var = ELSA_w6_weig$weight, missing_values = c(-1))
ggplot(ELSA_w6_weig, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency") + xlim(30,135)
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `recode -1 "Not Applicable" into 996`
* `NA into 999`

**R script:**

```{r harmo6}
ELSA_ds_w6 <- tibble(id=ELSA_w6_weig$id)
ELSA_ds_w6$weight <- car::recode(ELSA_w6_weig$weight, "NA=999; -1=996")
ELSA_ds_w6$weight <- labelled(ELSA_ds_w6$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript6, echo=F}
Continuous_summary(var = ELSA_ds_w6$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ELSA_ds_w6, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency") + xlim(30,135)
```

#### Validation
It is going to check if weight and heigth values are correct by computing BMI with the formula: weight/height^2. 
```{r crosstabulation6, echo=F}
opal.assign.table.tibble(o, 'ELSA_w6_whb','ELSA.wave_6_elsa_nurse_data_v2',
                         variables=list('weight', 'HTVAL', 'BMIVAL'), missings = TRUE)
ELSA_w6_whb <- opal.execute(o,'ELSA_w6_whb')

compute_bmi <- function(Heigth, Weigth, miss, heigth_m = TRUE){
  Heigth <- car::recode(Heigth, "NA=999")
  Weigth <- car::recode(Weigth, "NA=999")
  
  if(heigth_m){
    bmi <- Weigth/(Heigth^2)
    
    bmi[Heigth==999] <- 999
    bmi[Heigth==miss[1]] <- Heigth[Heigth==miss[1]]
    
    bmi[Weigth==999] <- 999
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[1]]

  } else {
    bmi <- Weigth/((Heigth/100)^2)
    
    bmi[Heigth==999] <- 999
    bmi[Heigth==miss[1]] <- Heigth[Heigth==miss[1]]
    
    bmi[Weigth==999] <- 999
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[1]]
  }
  bmi
}

ELSA_w6_whb$bmi_c <- compute_bmi(Heigth = ELSA_w6_whb$HTVAL, 
                        Weigth = ELSA_w6_whb$weight, miss = c(-1), 
                        heigth_m = FALSE)

ELSA_w6_whb$proff <- ELSA_w6_whb$BMIVAL - ELSA_w6_whb$bmi_c

head(ELSA_w6_whb[which(ELSA_w6_whb$proff>2 | ELSA_w6_whb$proff<(-2)),])

bmi_na <- which(ELSA_w6_whb$BMIVAL==(-1) & ELSA_w6_whb$HTVAL!=(-1) & ELSA_w6_whb$weight!=(-1))
weig_na <- which(ELSA_w6_whb$BMIVAL!=(-1) & ELSA_w6_whb$weight==(-1))
```

There are 66 individuals that the bmi value is different than the computed by the formula. Those discrepancies are because either bmi is missing (-1) or weight or height is missing (-1). Therefore, it is going to be used this values to complet the variables bmi, weight and height. 

```{r crosstabulation6_2, echo=F}
ELSA_ds_w6$weight[weig_na] <- ELSA_w6_whb$BMIVAL[weig_na] * (ELSA_w6_whb$HTVAL[weig_na]/100)^2

```

```{r importOPAL6, echo=F}
#val_labels(ELSA_ds_w6) <- NULL
#opal.assign.data(o, 'ELSA_ds_w6', ELSA_ds_w6)
#opal.symbol_import(o,'ELSA_ds_w6', project='_Athlos_Harm_Dataset')
```



```{r,  echo=FALSE}
weight <- ELSA_ds_w6
save(weight, file = paste0(datafolder, "/w6/weight.RData"))
```



### Wave 7

Not found.



```{r closeRsession, echo=FALSE}
opal.logout(o)
```


#### Quality estimation
No coments.
