==> ELSA/6_Physical_measures/7_Grip_Strength.Rmd <==
---
title: "Grip strength"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `grip`
* Variable label: `Grip strength in kg`
* Variable description: `Grip strength`
* Domain: `Physical measures`
* Value type: `continuous`

* Comments: `According to a systematic review (Roberts et al, 2011), there are some inconsistencies in the use of maximum or mean grip strength as a summary measure which limit comparison of results between epidemiological studies. With multiple attempts, the maximum grip strength will be greater than the mean value. 
Even though for right handed people, there is a 10% stronger grip in the dominant hand, this is not true for left handed or ambidextrous people, whose grip strength are more or less equal in both hands (Roberts et al, 2011). 
In light of the above, the harmonised variable "grip" is proposed to be the highest score (to the nearest kg) of all hand grip measurements in each study, regardless of hand dominance. This is following a protocol suggested by Roberts et al (2011), based on the American Society of Hand Therapists (ASHT) protocol.`

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Grip strength"
```

# Data process


```{r, dataDirectory, echo=FALSE}
datafolder <- "../RData"
```


## ELSA

### Wave 1 

There is no variable. 






### Wave 2 

#### Study-specific variable description


| **Name** | `mmgsd1`| `mmgsd2`| `mmgsd3`| 
|-|-|-|-|
| **Label** | `Grip strength: 1st measurement dominant hand (kg)` | `Grip strength: 2nd measurement dominant hand (kg)` | `Grip strength: 3rd measurement dominant hand (kg)` |
| **Table name** | `wave_2_nurse_data_v2` |`wave_2_nurse_data_v2` |`wave_2_nurse_data_v2` |
| **Categories** | `continuous` |`continuous` |`continuous` |
| **Missing values** | `-1 = Not Applicable` | `-1 = Not Applicable` | `-1 = Not Applicable` | 
| | `-8 = Dont know` |  `-8 = Dont know` |  `-8 = Dont know` | 
| | `-9 = No answer/Refusal` | `-9 = No answer/Refusal` | `-9 = No answer/Refusal` | 

| **Name** | `mmgsn1`| `mmgsn2`| `mmgsn3`| 
|-|-|-|-|
| **Label** | `Grip strength: 1st measurement non-dominant hand (kg)` | `Grip strength: 2nd measurement non-dominant hand (kg)` | `Grip strength: 3rd measurement non-dominant hand (kg)` |
| **Table name** | `wave_2_nurse_data_v2` |`wave_2_nurse_data_v2` |`wave_2_nurse_data_v2` |
| **Categories** | `continuous` |`continuous` |`continuous` |
| **Missing values** | `-1 = Not Applicable` | `-1 = Not Applicable` | `-1 = Not Applicable` | 
| | `-8 = Dont know` |  `-8 = Dont know` |  `-8 = Dont know` | 
| | `-9 = No answer/Refusal` | `-9 = No answer/Refusal` | `-9 = No answer/Refusal` | 
| **Description** |  | | |


```{r assign2, echo=FALSE}
opal.assign.table.tibble(o, 'ELSA_gs','ELSA.wave_2_nurse_data_v2',
                         variables=list('mmgsd1', 'mmgsd2', 'mmgsd3', 'mmgsn1', 'mmgsn2', 'mmgsn3'), missings = TRUE)
ELSA_gs <- opal.execute(o,'ELSA_gs')
```

```{r local2, echo=FALSE}
vari <- ELSA_gs$mmgsd1
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsd1)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsd2
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsd2)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsd3
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsd3)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)


vari <- ELSA_gs$mmgsn1
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsn1)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsn2
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsn2)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsn3
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsn3)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `-9 into 997`
* `-8 into 998`
* `-1 into 996`
* `NA into 999`

**R script:**

```{r harmo2}
ELSA_ds_w2 <- tibble(id=ELSA_gs$id)
ELSA_ds_w2$grip <- apply(ELSA_gs[,2:7], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
ELSA_ds_w2$grip <- car::recode(ELSA_ds_w2$grip, "-Inf=999; -1=996; -8=998; -9=997")
ELSA_ds_w2$grip <- as.integer(ELSA_ds_w2$grip)
ELSA_ds_w2$grip <- labelled(ELSA_ds_w2$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=FALSE}
vari <- ELSA_ds_w2$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ELSA_ds_w2, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation2, echo=FALSE}
```

```{r importOPAL2, echo=FALSE}
#val_labels(ELSA_ds_w2) <- NULL
#opal.assign.data(o, 'ELSA_ds_w2', ELSA_ds_w2)
#opal.symbol_import(o,'ELSA_ds_w2', project='_Athlos_Harm_Dataset')
label(ELSA_ds_w2$grip) <- ds_label
grip <- ELSA_ds_w2
save(grip, file = "../RData/w2/grip.RData")
rm(grip)
```







### Wave 3

There is no variable.



### Wave 4 

#### Study-specific variable description

| **Name** | `mmgsd1`| `mmgsd2`| `mmgsd3`| 
|-|-|-|-|
| **Label** | `Grip strength: 1st measurement dominant hand (kg)` | `Grip strength: 2nd measurement dominant hand (kg)` | `Grip strength: 3rd measurement dominant hand (kg)` |
| **Table name** | `wave_4_nurse_data` |`wave_4_nurse_data` |`wave_4_nurse_data` |
| **Categories** | `continuous` |`continuous` |`continuous` |
| **Missing values** | `-1 = Not Applicable` | `-1 = Not Applicable` | `-1 = Not Applicable` | 
| | `-8 = Dont know` |  `-8 = Dont know` |  `-8 = Dont know` | 
| | `-9 = No answer/Refusal` | `-9 = No answer/Refusal` | `-9 = No answer/Refusal` | 

| **Name** | `mmgsn1`| `mmgsn2`| `mmgsn3`| 
|-|-|-|-|
| **Label** | `Grip strength: 1st measurement non-dominant hand (kg)` | `Grip strength: 2nd measurement non-dominant hand (kg)` | `Grip strength: 3rd measurement non-dominant hand (kg)` |
| **Table name** | `wave_4_nurse_data` |`wave_4_nurse_data` |`wave_4_nurse_data` |
| **Categories** | `continuous` |`continuous` |`continuous` |
| **Missing values** | `-1 = Not Applicable` | `-1 = Not Applicable` | `-1 = Not Applicable` | 
| | `-8 = Dont know` |  `-8 = Dont know` |  `-8 = Dont know` | 
| | `-9 = Refusal` | `-9 = Refusal` | `-9 = Refusal` | 
| | `99 = Measurement not obtained` | `99 = Measurement not obtained` | `99 = Measurement not obtained` | 
| **Description** |  | | |

```{r assign4, echo=FALSE}
opal.assign.table.tibble(o, 'ELSA_gs','ELSA.wave_4_nurse_data',
                         variables=list('mmgsd1', 'mmgsd2', 'mmgsd3', 'mmgsn1', 'mmgsn2', 'mmgsn3'), missings = TRUE)
ELSA_gs <- opal.execute(o,'ELSA_gs')
```

```{r local4, echo=FALSE}
vari <- ELSA_gs$mmgsd1
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9, 99))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsd1)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsd2
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9, 99))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsd2)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsd3
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9, 99))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsd3)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)


vari <- ELSA_gs$mmgsn1
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9, 99))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsn1)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsn2
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9, 99))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsn2)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsn3
kable(Continuous_summary(var = vari, missing_values = c(-1, -8, -9, 99))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -8, -9, 99))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsn3)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `-9 into 997`
* `-8 into 998`
* `-1 into 996`
* `99 into 999`
* `NA into 999`

**R script:**


```{r harmo4}
ELSA_gs$mmgsd1 <- car::recode(ELSA_gs$mmgsd1, "99=-2")
ELSA_gs$mmgsd2 <- car::recode(ELSA_gs$mmgsd2, "99=-2")
ELSA_gs$mmgsd3 <- car::recode(ELSA_gs$mmgsd3, "99=-2")
ELSA_gs$mmgsn1 <- car::recode(ELSA_gs$mmgsn1, "99=-2")
ELSA_gs$mmgsn2 <- car::recode(ELSA_gs$mmgsn2, "99=-2")
ELSA_gs$mmgsn3 <- car::recode(ELSA_gs$mmgsn3, "99=-2")

ELSA_ds_w4 <- tibble(id=ELSA_gs$id)
ELSA_ds_w4$grip <- apply(ELSA_gs[,2:7], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
ELSA_ds_w4$grip <- car::recode(ELSA_ds_w4$grip, "-Inf=999; -1=996; -8=998; -9=997; -2=999")
ELSA_ds_w4$grip <- as.integer(ELSA_ds_w4$grip)
ELSA_ds_w4$grip <- labelled(ELSA_ds_w4$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript4, echo=FALSE}
vari <- ELSA_ds_w4$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ELSA_ds_w4, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation4, echo=FALSE}
```

```{r importOPAL4, echo=FALSE}
label(ELSA_ds_w4$grip) <- ds_label
grip <- ELSA_ds_w4
save(grip, file = "../RData/w4/grip.RData")
rm(grip)
```






### Wave 5

There is no variable.




### Wave 6 

#### Study-specific variable description

| **Name** | `mmgsd1`| `mmgsd2`| `mmgsd3`| 
|-|-|-|-|
| **Label** | `Grip strength: 1st measurement dominant hand (kg)` | `Grip strength: 2nd measurement dominant hand (kg)` | `Grip strength: 3rd measurement dominant hand (kg)` |
| **Table name** | `wave_6_elsa_nurse_data_v2` |`wave_6_elsa_nurse_data_v2` |`wave_6_elsa_nurse_data_v2` |
| **Categories** | `continuous` |`continuous` |`continuous` |
| **Missing values** | `-1 = Not Applicable` | `-1 = Not Applicable` | `-1 = Not Applicable` | 

| **Name** | `mmgsn1`| `mmgsn2`| `mmgsn3`| 
|-|-|-|-|
| **Label** | `Grip strength: 1st measurement non-dominant hand (kg)` | `Grip strength: 2nd measurement non-dominant hand (kg)` | `Grip strength: 3rd measurement non-dominant hand (kg)` |
| **Table name** | `wave_6_elsa_nurse_data_v2` |`wave_6_elsa_nurse_data_v2` |`wave_6_elsa_nurse_data_v2` |
| **Categories** | `continuous` |`continuous` |`continuous` |
| **Missing values** | `-1 = Not Applicable` | `-1 = Not Applicable` | `-1 = Not Applicable` | 
| **Description** |  | | |


```{r assign6, echo=FALSE}
opal.assign.table.tibble(o, 'ELSA_gs','ELSA.wave_6_elsa_nurse_data_v2',
                         variables=list('mmgsd1', 'mmgsd2', 'mmgsd3', 'mmgsn1', 'mmgsn2', 'mmgsn3'), missings = TRUE)
ELSA_gs <- opal.execute(o,'ELSA_gs')
```

```{r local6, echo=FALSE}
vari <- ELSA_gs$mmgsd1
kable(Continuous_summary(var = vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsd1)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsd2
kable(Continuous_summary(var = vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsd2)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsd3
kable(Continuous_summary(var = vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsd3)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)


vari <- ELSA_gs$mmgsn1
kable(Continuous_summary(var = vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsn1)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsn2
kable(Continuous_summary(var = vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsn2)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)

vari <- ELSA_gs$mmgsn3
kable(Continuous_summary(var = vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[3])
kable(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-1, -2, -3, -4, -5, -6 ,-7 ,-8, -9))$summary, caption = "Summary")
ggplot(ELSA_gs, aes(mmgsn3)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100)
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `-2:-9 into 999`
* `-1 into 996`
* `NA into 999`

**R script:**


```{r harmo6}
ELSA_ds_w6 <- tibble(id=ELSA_gs$id)
ELSA_ds_w6$grip <- apply(ELSA_gs[,2:7], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
ELSA_ds_w6$grip <- car::recode(ELSA_ds_w6$grip, "-Inf=999; -1=996; -2:-9=999")
ELSA_ds_w6$grip <- as.integer(ELSA_ds_w6$grip)
ELSA_ds_w6$grip <- labelled(ELSA_ds_w6$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript6, echo=FALSE}
vari <- ELSA_ds_w6$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ELSA_ds_w6, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation6, echo=FALSE}
```


```{r importOPAL6, echo=FALSE}
label(ELSA_ds_w6$grip) <- ds_label
grip <- ELSA_ds_w6
save(grip, file = "../RData/w6/grip.RData")
rm(grip)
```

### Wave 7

There is no variable.




# Quality estimation
No comments.



```{r closeRsession, echo=FALSE}
opal.logout(o)
```
