==> ELSA/6_Physical_measures/11_Diastolic.Rmd <==
---
title: "Diastolic blood pressure"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("../../setup_iago.r")
```
<!-- ########################################################## --> 


# Description of DataSchema variable

The description of harmonised variable is the following:


* Short name: `dbp`
* Variable label: `Diastolic blood pressure of participant at each wave`
* Variable description: `Means of all diastolic blood pressure measures.`
* Domain: `Physical measures `
* Value type: `Continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing




<!-- ########################################################## --> 

# Data process






```{r, dataDirectory, echo=FALSE}
datafolder <- "../RData"
```

## England


### Wave 1 

There is no variable. 






### Wave 2

#### Study-specific variable description

The study-specific variable elected to be harmonised is:


* Name: `dias1`
* Label: `BP: 1st diastolic reading (mmHg)`
* Type of variable: `countinous`
* Missings
    + `-9 = No answer/Refusal`
    + `-8 = Dont know`
    + `-3 = Error reading`
    + `-2 = No reading obtained`
    + `-1 = Not Applicable`

* Name: `dias2`
* Label: `BP: 2nd diastolic reading (mmHg)`
* Type of variable: `countinous`
* Missings
    + `-9 = No answer/Refusal`
    + `-8 = Dont know`
    + `-3 = Error reading`
    + `-2 = No reading obtained`
    + `-1 = Not Applicable`

* Name: `dias3`
* Label: `BP: 3rd diastolic reading (mmHg)`
* Type of variable: `countinous`
* Missings
    + `-9 = No answer/Refusal`
    + `-8 = Dont know`
    + `-3 = Error reading`
    + `-2 = No reading obtained`
    + `-1 = Not Applicable`

```{r assign2, echo=F}
opal.assign.table.tibble(o, 'ELSA_w2_dbp','ELSA.wave_2_nurse_data_v2',
                         variables=list('dias1', 'dias2', 'dias3'), missings = TRUE)
```

```{r local2, echo=F}
ELSA_w2_dbp <- opal.execute(o,'ELSA_w2_dbp')

Continuous_summary(var = ELSA_w2_dbp$dias1, missing_values = c(-9, -8, -3, -2, -1))
ggplot(filter(ELSA_w2_dbp,!ELSA_w2_dbp$dias1 %in% c(-9, -8, -3, -2, -1)), aes(dias1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ELSA_w2_dbp$dias2, missing_values = c(-9, -8, -3, -2, -1))
ggplot(filter(ELSA_w2_dbp,!ELSA_w2_dbp$dias2 %in% c(-9, -8, -3, -2, -1)), aes(dias2)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ELSA_w2_dbp$dias3, missing_values = c(-9, -8, -3, -2, -1))
ggplot(filter(ELSA_w2_dbp,!ELSA_w2_dbp$dias3 %in% c(-9, -8, -3, -2, -1)), aes(dias3)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first, second and third measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `recode -9 "No answer/Refusal" into 995`
* `recode -8 "Dont know" into 998`
* `recode -3 "Error reading" into 991`
* `recode -2 "No reading obtained" into 999`
* `recode -1 "Not Applicable" into 996`
* `NA into 999`

**R script:**

```{r harmo2}
ELSA_w2_dbp$mean_dbp <- apply(X = ELSA_w2_dbp[,c(2,3)], MARGIN = 1, FUN = mean)

### -2 only in the first one
id2_only01 <- which(ELSA_w2_dbp$dias1==-2 & ELSA_w2_dbp$dias2!=-2 & ELSA_w2_dbp$dias3!=-2)

ELSA_w2_dbp$mean_dbp[id2_only01] <- apply(X = ELSA_w2_dbp[id2_only01,c(3,4)], MARGIN = 1, FUN = mean)

### -2 only in the third one
id2_only03 <- which(ELSA_w2_dbp$dias1!=-2 & ELSA_w2_dbp$dias2!=-2 & ELSA_w2_dbp$dias3==-2)

ELSA_w2_dbp$mean_dbp[id2_only03] <- apply(X = ELSA_w2_dbp[id2_only03,c(2,3)], MARGIN = 1, FUN = mean)

### -2 in the first and second one
id2_01_02 <- which(ELSA_w2_dbp$dias1==-2 & ELSA_w2_dbp$dias2==-2 & ELSA_w2_dbp$dias3!=-2)

ELSA_w2_dbp$mean_dbp[id2_01_02] <- ELSA_w2_dbp$dias3[id2_01_02]

### -2 in the first and third one
id2_01_03 <- which(ELSA_w2_dbp$dias1==-2 & ELSA_w2_dbp$dias2!=-2 & ELSA_w2_dbp$dias3==-2)

ELSA_w2_dbp$mean_dbp[id2_01_03] <- ELSA_w2_dbp$dias2[id2_01_03]

### -2 in the second and third one
id2_02_03 <- which(ELSA_w2_dbp$dias1!=-2 & ELSA_w2_dbp$dias2==-2 & ELSA_w2_dbp$dias3==-2)

ELSA_w2_dbp$mean_dbp[id2_02_03] <- ELSA_w2_dbp$dias1[id2_02_03]





ELSA_ds_w2 <- tibble(id=ELSA_w2_dbp$id)
ELSA_ds_w2$dbp <- car::recode(ELSA_w2_dbp$mean_dbp, "NA=999; -9=995; -8=998; -3=991; -2=999; -1=996")
ELSA_ds_w2$dbp <- labelled(ELSA_ds_w2$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=F}
Continuous_summary(var = ELSA_ds_w2$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ELSA_ds_w2, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation2, echo=F}
```

```{r importOPAL2}
#val_labels(ELSA_ds_w2) <- NULL
#opal.assign.data(o, 'ELSA_ds_w2', ELSA_ds_w2)
#opal.symbol_import(o,'ELSA_ds_w2', project='_Athlos_Harm_Dataset')
```







### Wave 3

There is no variable.



### Wave 4

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `dias1`
* Label: `BP: 1st diastolic reading (mmHg)`
* Type of variable: `countinous`
* Missings
    + `-1 = Not Applicable`
    + `999 = Reading not obtained`

* Name: `dias2`
* Label: `BP: 2nd diastolic reading (mmHg)`
* Type of variable: `countinous`
* Missings
    + `-1 = Not Applicable`
    + `999 = Reading not obtained`

* Name: `dias3`
* Label: `BP: 3rd diastolic reading (mmHg)`
* Type of variable: `countinous`
* Missings
    + `-1 = Not Applicable`
    + `999 = Reading not obtained`

```{r assign4, echo=F}
opal.assign.table.tibble(o, 'ELSA_w4_dbp','ELSA.wave_4_nurse_data',
                         variables=list('dias1', 'dias2', 'dias3'), missings = TRUE)
```

```{r local4, echo=F}
ELSA_w4_dbp <- opal.execute(o,'ELSA_w4_dbp')

Continuous_summary(var = ELSA_w4_dbp$dias1, missing_values = c(-1, 999))
ggplot(filter(ELSA_w4_dbp,!ELSA_w4_dbp$dias1 %in% c(-1,999)), aes(dias1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ELSA_w4_dbp$dias2, missing_values = c(-1, 999))
ggplot(filter(ELSA_w4_dbp,!ELSA_w4_dbp$dias2 %in% c(-1,999)), aes(dias2)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ELSA_w4_dbp$dias3, missing_values = c(-1, 999))
ggplot(filter(ELSA_w4_dbp,!ELSA_w4_dbp$dias3 %in% c(-1,999)), aes(dias3)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first, second and third measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `keep "Reading not obtained" into 999`
* `recode -1 "Not Applicable" into 996`
* `NA into 999`

**R script:**

```{r harmo4}
ELSA_w4_dbp$dias1 <- car::recode(ELSA_w4_dbp$dias1, "999=-2")
ELSA_w4_dbp$dias2 <- car::recode(ELSA_w4_dbp$dias2, "999=-2")
ELSA_w4_dbp$dias3 <- car::recode(ELSA_w4_dbp$dias3, "999=-2")

ELSA_w4_dbp$mean_dbp <- apply(X = ELSA_w4_dbp[,c(2,3)], MARGIN = 1, FUN = mean)

### -2 only in the first one
id2_only01 <- which(ELSA_w4_dbp$dias1==-2 & ELSA_w4_dbp$dias2!=-2 & ELSA_w4_dbp$dias3!=-2)

ELSA_w4_dbp$mean_dbp[id2_only01] <- apply(X = ELSA_w4_dbp[id2_only01,c(3,4)], MARGIN = 1, FUN = mean)

### -2 only in the second one
id2_only02 <- which(ELSA_w4_dbp$dias1!=-2 & ELSA_w4_dbp$dias2==-2 & ELSA_w4_dbp$dias3!=-2)

ELSA_w4_dbp$mean_dbp[id2_only02] <- apply(X = ELSA_w4_dbp[id2_only02,c(2,4)], MARGIN = 1, FUN = mean)

### -2 only in the third one
id2_only03 <- which(ELSA_w4_dbp$dias1!=-2 & ELSA_w4_dbp$dias2!=-2 & ELSA_w4_dbp$dias3==-2)

ELSA_w4_dbp$mean_dbp[id2_only03] <- apply(X = ELSA_w4_dbp[id2_only03,c(2,3)], MARGIN = 1, FUN = mean)

### -2 in the first and second one
id2_01_02 <- which(ELSA_w4_dbp$dias1==-2 & ELSA_w4_dbp$dias2==-2 & ELSA_w4_dbp$dias3!=-2)

ELSA_w4_dbp$mean_dbp[id2_01_02] <- ELSA_w4_dbp$dias3[id2_01_02]

### -2 in the second and third one
id2_02_03 <- which(ELSA_w4_dbp$dias1!=-2 & ELSA_w4_dbp$dias2==-2 & ELSA_w4_dbp$dias3==-2)

ELSA_w4_dbp$mean_dbp[id2_02_03] <- ELSA_w4_dbp$dias1[id2_02_03]





ELSA_ds_w4 <- tibble(id=ELSA_w4_dbp$id)
ELSA_ds_w4$dbp <- car::recode(ELSA_w4_dbp$mean_dbp, "NA=999; -2=999; -1=996")
ELSA_ds_w4$dbp <- labelled(ELSA_ds_w4$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript4, echo=F}
Continuous_summary(var = ELSA_ds_w4$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ELSA_ds_w4, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation4, echo=F}
```

```{r importOPAL4, echo=F}
#val_labels(ELSA_ds_w4) <- NULL
#opal.assign.data(o, 'ELSA_ds_w4', ELSA_ds_w4)
#opal.symbol_import(o,'ELSA_ds_w4', project='_Athlos_Harm_Dataset')
```






### Wave 5

There is no variable.


### Wave 6

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `dias1`
* Label: `BP: 1st diastolic reading (mmHg)`
* Type of variable: `countinous`
* Missings
    + `-1 = Not Applicable`
    + `999 = Reading not obtained`

* Name: `dias2`
* Label: `BP: 2nd diastolic reading (mmHg)`
* Type of variable: `countinous`
* Missings
    + `-1 = Not Applicable`
    + `999 = Reading not obtained`

* Name: `dias3`
* Label: `BP: 3rd diastolic reading (mmHg)`
* Type of variable: `countinous`
* Missings
    + `-1 = Not Applicable`
    + `999 = Reading not obtained`


```{r assign6, echo=F}
opal.assign.table.tibble(o, 'ELSA_w6_dbp','ELSA.wave_6_elsa_nurse_data_v2',
                         variables=list('dias1', 'dias2', 'dias3'), missings = TRUE)
```

```{r local6, echo=F}
ELSA_w6_dbp <- opal.execute(o,'ELSA_w6_dbp')

Continuous_summary(var = ELSA_w6_dbp$dias1, missing_values = c(-1, 999))
ggplot(filter(ELSA_w6_dbp,!ELSA_w6_dbp$dias1 %in% c(-1,999)), aes(dias1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ELSA_w6_dbp$dias2, missing_values = c(-1, 999))
ggplot(filter(ELSA_w6_dbp,!ELSA_w6_dbp$dias2 %in% c(-1,999)), aes(dias2)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ELSA_w6_dbp$dias3, missing_values = c(-1, 999))
ggplot(filter(ELSA_w6_dbp,!ELSA_w6_dbp$dias3 %in% c(-1,999)), aes(dias3)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first, second and third measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `keep "Reading not obtained" into 999`
* `recode -1 "Not Applicable" into 996`
* `NA into 999`

**R script:**

```{r harmo6}
ELSA_w6_dbp$dias1 <- car::recode(ELSA_w6_dbp$dias1, "999=-2")
ELSA_w6_dbp$dias2 <- car::recode(ELSA_w6_dbp$dias2, "999=-2")
ELSA_w6_dbp$dias3 <- car::recode(ELSA_w6_dbp$dias3, "999=-2")

ELSA_w6_dbp$mean_dbp <- apply(X = ELSA_w6_dbp[,c(2,3)], MARGIN = 1, FUN = mean)

### -2 only in the second one
id2_only02 <- which(ELSA_w6_dbp$dias1!=-2 & ELSA_w6_dbp$dias2==-2 & ELSA_w6_dbp$dias3!=-2)

ELSA_w6_dbp$mean_dbp[id2_only02] <- apply(X = ELSA_w6_dbp[id2_only02,c(2,4)], MARGIN = 1, FUN = mean)

### -2 only in the third one
id2_only03 <- which(ELSA_w6_dbp$dias1!=-2 & ELSA_w6_dbp$dias2!=-2 & ELSA_w6_dbp$dias3==-2)

ELSA_w6_dbp$mean_dbp[id2_only03] <- apply(X = ELSA_w6_dbp[id2_only03,c(2,3)], MARGIN = 1, FUN = mean)

### -2 in the second and third one
id2_02_03 <- which(ELSA_w6_dbp$dias1!=-2 & ELSA_w6_dbp$dias2==-2 & ELSA_w6_dbp$dias3==-2)

ELSA_w6_dbp$mean_dbp[id2_02_03] <- ELSA_w6_dbp$dias1[id2_02_03]


ELSA_ds_w6 <- tibble(id=ELSA_w6_dbp$id)
ELSA_ds_w6$dbp <- car::recode(ELSA_w6_dbp$mean_dbp, "NA=999; -2=999; -1=996")
ELSA_ds_w6$dbp <- labelled(ELSA_ds_w6$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript6, echo=F}
Continuous_summary(var = ELSA_ds_w6$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ELSA_ds_w6, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation6, echo=F}
```

```{r importOPAL6, echo=F}
#val_labels(ELSA_ds_w6) <- NULL
#opal.assign.data(o, 'ELSA_ds_w6', ELSA_ds_w6)
#opal.symbol_import(o,'ELSA_ds_w6', project='_Athlos_Harm_Dataset')
```




### Wave 7

There is no variable.


# Quality estimation

Diastolic  bp was calculated as the mean from the 3 available variables.


```{r save, echo=F}
s <- list(ELSA_ds_w2, ELSA_ds_w4, ELSA_ds_w6)
for(i in 1:3)
{
  dbp <- s[[i]]
  save(dbp, file = paste0(datafolder, "/",paste0("w",2*i),"/dbp.RData"))
}

```
