==> ELSA/6_Physical_measures/12_ah.Rmd <==
---
title: "Arterial hypertension"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---


```{r setup, include=FALSE}
source("../../setup_iago.r")
```

# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `ah`
* Variable label: `Arterial hypertension`
* Variable description: `Arterial hypertension. Means of all diastolic blood pressure measures.`
* Domain: `Physical Measures`
* Value type: `Categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | no hypertension
1        | hypertension

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done 
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Arterial hypertension"
ds_label_all <- "Arterial hypertension. Means of all diastolic blood pressure measures."
```
<!-- ########################################################## --> 

# Data process


```{r, dataDirectory, echo=FALSE}
datafolder <- "../RData"
```


## ELSA

### Wave 1 

There is no variable. 






### Wave 2 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |


```{r assign2, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ELSA/RData/w2/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ELSA/RData/w2/dbp.RData")
ELSA1_ah <- merge(sbp, dbp, by = "id")
```

```{r local2, echo=FALSE}
vari <- ELSA1_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ELSA1_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- ELSA1_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ELSA1_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp>= 140 into 1`
  + `sbp< 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo2}
ELSA1_ah$dh <- car::recode(ELSA1_ah$dbp, "0:89.99999999=0; 90:900=1")
ELSA1_ah$sh <- car::recode(ELSA1_ah$sbp, "0:139.9999999=0; 140:900=1")

ELSA_ds_w2 <- tibble(id=ELSA1_ah$id)
ELSA_ds_w2$ah <- NA
ELSA_ds_w2$ah[which(ELSA1_ah$dh==0 & ELSA1_ah$sh==0)] <- 0
ELSA_ds_w2$ah[which(ELSA1_ah$dh==1 & ELSA1_ah$sh==0)] <- 1
ELSA_ds_w2$ah[which(ELSA1_ah$dh==0 & ELSA1_ah$sh==1)] <- 1
ELSA_ds_w2$ah[which(ELSA1_ah$dh==1 & ELSA1_ah$sh==1)] <- 1
ELSA_ds_w2$ah[which(ELSA1_ah$dh==999 & ELSA1_ah$sh==999)] <- 999
ELSA_ds_w2$ah[which(ELSA1_ah$dh==996 & ELSA1_ah$sh==996)] <- 996

ELSA_ds_w2$ah[which(ELSA1_ah$dh==1 & ELSA1_ah$sh==999)] <- 1
ELSA_ds_w2$ah[which(ELSA1_ah$dh==999 & ELSA1_ah$sh==0)] <- 999

ELSA_ds_w2$ah <- labelled(ELSA_ds_w2$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=FALSE}
vari <- ELSA_ds_w2$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(ELSA_ds_w2, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation2, echo=FALSE}
```

```{r importOPAL2, echo=FALSE}
#val_labels(ELSA_ds_w2) <- NULL
#opal.assign.data(o, 'ELSA_ds_w2', ELSA_ds_w2)
#opal.symbol_import(o,'ELSA_ds_w2', project='_Athlos_Harm_Dataset')
label(ELSA_ds_w2$ah) <- ds_label_all
ah <- ELSA_ds_w2
save(ah, file = "../RData/w2/ah.RData")
rm(ah)
```







### Wave 3

There is no variable.



### Wave 4 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |

```{r assign4, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ELSA/RData/w4/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ELSA/RData/w4/dbp.RData")
ELSA2_ah <- merge(sbp, dbp, by = "id")
```

```{r local4, echo=FALSE}
vari <- ELSA2_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ELSA2_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- ELSA2_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ELSA2_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp>= 140 into 1`
  + `sbp< 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo4}
ELSA2_ah$dh <- car::recode(ELSA2_ah$dbp, "0:89.99999999=0; 90:900=1")
ELSA2_ah$sh <- car::recode(ELSA2_ah$sbp, "0:139.9999999=0; 140:900=1")

ELSA_ds_w4 <- tibble(id=ELSA2_ah$id)
ELSA_ds_w4$ah <- NA
ELSA_ds_w4$ah[which(ELSA2_ah$dh==0 & ELSA2_ah$sh==0)] <- 0
ELSA_ds_w4$ah[which(ELSA2_ah$dh==1 & ELSA2_ah$sh==0)] <- 1
ELSA_ds_w4$ah[which(ELSA2_ah$dh==0 & ELSA2_ah$sh==1)] <- 1
ELSA_ds_w4$ah[which(ELSA2_ah$dh==1 & ELSA2_ah$sh==1)] <- 1
ELSA_ds_w4$ah[which(ELSA2_ah$dh==999 & ELSA2_ah$sh==999)] <- 999
ELSA_ds_w4$ah[which(ELSA2_ah$dh==996 & ELSA2_ah$sh==996)] <- 996

ELSA_ds_w4$ah[which(ELSA2_ah$dh==0 & ELSA2_ah$sh==999)] <- 999

ELSA_ds_w4$ah <- labelled(ELSA_ds_w4$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript4, echo=FALSE}
vari <- ELSA_ds_w4$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(ELSA_ds_w4, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation4, echo=FALSE}
```

```{r importOPAL4, echo=FALSE}
label(ELSA_ds_w4$ah) <- ds_label_all
ah <- ELSA_ds_w4
save(ah, file = "../RData/w4/ah.RData")
rm(ah)
```






### Wave 5

There is no variable.




### Wave 6 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |

```{r assign6, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ELSA/RData/w6/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ELSA/RData/w6/dbp.RData")
ELSA3_ah <- merge(sbp, dbp, by = "id")
```

```{r local6, echo=FALSE}
vari <- ELSA3_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ELSA3_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- ELSA3_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ELSA3_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp>= 140 into 1`
  + `sbp< 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo6}
ELSA3_ah$dh <- car::recode(ELSA3_ah$dbp, "0:89.99999999=0; 90:900=1")
ELSA3_ah$sh <- car::recode(ELSA3_ah$sbp, "0:139.9999999=0; 140:900=1")

ELSA_ds_w6 <- tibble(id=ELSA3_ah$id)
ELSA_ds_w6$ah <- NA
ELSA_ds_w6$ah[which(ELSA3_ah$dh==0 & ELSA3_ah$sh==0)] <- 0
ELSA_ds_w6$ah[which(ELSA3_ah$dh==1 & ELSA3_ah$sh==0)] <- 1
ELSA_ds_w6$ah[which(ELSA3_ah$dh==0 & ELSA3_ah$sh==1)] <- 1
ELSA_ds_w6$ah[which(ELSA3_ah$dh==1 & ELSA3_ah$sh==1)] <- 1
ELSA_ds_w6$ah[which(ELSA3_ah$dh==999 & ELSA3_ah$sh==999)] <- 999
ELSA_ds_w6$ah[which(ELSA3_ah$dh==996 & ELSA3_ah$sh==996)] <- 996
ELSA_ds_w6$ah <- labelled(ELSA_ds_w6$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript6, echo=FALSE}
vari <- ELSA_ds_w6$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(ELSA_ds_w6, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation6, echo=FALSE}
```

```{r importOPAL6, echo=FALSE}
label(ELSA_ds_w6$ah) <- ds_label_all
ah <- ELSA_ds_w6
save(ah, file = "../RData/w6/ah.RData")
rm(ah)
```



### Wave 7

There is no variable.


# Quality estimation

No comments.


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
